+++ src/main/java/com/magneticraft2/common/systems/heat/HeatStorage.java
@@ -2,21 +2,21 @@ package com.magneticraft2.common.systems.heat;
 
 public class HeatStorage implements IHeatStorage{
 
-    protected double heat;
-    protected double capacity;
-    protected double maxReceive;
-    protected double maxExtract;
+    protected int heat;
+    protected int capacity;
+    protected int maxReceive;
+    protected int maxExtract;
 
-    public HeatStorage(double capacity) {
+    public HeatStorage(int capacity) {
         this(capacity,capacity,capacity,0);
     }
-    public HeatStorage(double capacity, double maxTransfer){
+    public HeatStorage(int capacity, int maxTransfer){
         this(capacity,maxTransfer,maxTransfer,0);
     }
-    public HeatStorage(double capacity, double maxReceive, double maxExtract){
+    public HeatStorage(int capacity, int maxReceive, int maxExtract){
         this(capacity,maxReceive,maxExtract,0);
     }
-    public HeatStorage(double capacity, double maxReceive, double maxExtract, double heat){
+    public HeatStorage(int capacity, int maxReceive, int maxExtract, int heat){
         this.capacity = capacity;
         this.maxReceive = maxReceive;
         this.maxExtract = maxExtract;
@@ -24,32 +24,32 @@ public class HeatStorage implements IHeatStorage{
     }
 
     @Override
-    public double receiveHeat(double maxReceive, boolean simulate) {
+    public int receiveHeat(int maxReceive, boolean simulate) {
         if (!canReceive())
             return 0;
-        double heatReceived = Math.min(capacity - heat, Math.min(this.maxReceive,maxReceive));
+        int heatReceived = Math.min(capacity - heat, Math.min(this.maxReceive,maxReceive));
         if (!simulate)
             heat += heatReceived;
         return heatReceived;
     }
 
     @Override
-    public double extractHeat(double maxExtract, boolean simulate) {
+    public int extractHeat(int maxExtract, boolean simulate) {
         if (!canReceive())
             return 0;
-        double heatExtracted = Math.min(heat, Math.min(this.maxExtract,maxExtract));
+        int heatExtracted = Math.min(heat, Math.min(this.maxExtract,maxExtract));
         if (!simulate)
             heat -= heatExtracted;
         return heatExtracted;
     }
 
     @Override
-    public double getHeatStored() {
+    public int getHeatStored() {
         return heat;
     }
 
     @Override
-    public double getMaxHeatStored() {
+    public int getMaxHeatStored() {
         return capacity;
     }
