+++ src/main/java/com/magneticraft2/common/systems/heat/IHeatStorage.java
@@ -2,10 +2,10 @@ package com.magneticraft2.common.systems.heat;
 
 public interface IHeatStorage {
 
-    double receiveHeat(double maxReceive, boolean simulate);
-    double extractHeat(double maxExtract, boolean simulate);
-    double getHeatStored();
-    double getMaxHeatStored();
+    int receiveHeat(int maxReceive, boolean simulate);
+    int extractHeat(int maxExtract, boolean simulate);
+    int getHeatStored();
+    int getMaxHeatStored();
     boolean canSend();
     boolean canReceive();
 }
